Add NavList component
